Rename ProcedureGuard to TerminationGuard and move to procedure api package
